success token

+++ seeds.js
@@ -7,13 +7,18 @@ const cyfer = bcryptjs.genSaltSync(8);
 
 const seedUsers = async () => {
     try {
-        await User.deleteMany()
-        const users = [
-            { name: "Owner of Market", username: "owner", email: "owner@mmarket.com", password: bcryptjs.hashSync(process.env.OWNER_PASSWORD, cyfer), verificated: true, status: "owner" },
-            { name: "Assembler", username: "assembler", email: "asm@we.com", password: bcryptjs.hashSync(process.env.OWNER_PASSWORD, cyfer), verificated: true, status: "owner" }
-        ]
-        await User.insertMany(users)
-        console.log("Seeded users successfully!");
+        const owner = await User.findOne({status: "owner"})
+        if (owner) {
+            await User.deleteMany()
+            const users = [
+                { name: "Owner of Market", username: "owner", email: "owner@mmarket.com", password: bcryptjs.hashSync(process.env.OWNER_PASSWORD, cyfer), verificated: true, status: "owner" },
+                { name: "Assembler", username: "assembler", email: "asm@we.com", password: bcryptjs.hashSync(process.env.OWNER_PASSWORD, cyfer), verificated: true, status: "owner" }
+            ]
+            await User.insertMany(users)
+            console.log("Seeded users successfully!");
+        } else {
+            console.log("Owners already found!");
+        }
     } catch (err) {
         console.log(`user seeding error ${err}`);
     }
